Update Template to latest version

+++ TemplateProject/TemplateProject/PiE/include/Camera4D.h
@@ -5,7 +5,7 @@
 class Camera4D :
 	public Camera
 {
-	Matrix<float, 5> viewMatrix4D = Matrix<float, 5>::Identity();
+	Matrix<double, 5> viewMatrix4D = Matrix<double, 5>::Identity();
 	Matrix4f viewMatrix3D = Matrix4f::Identity();
 	Vec4f pos;
 	float pitch4D = 0;
@@ -21,8 +21,8 @@ public:
 	const Vec3f getPos() const override;
 	const Vec4f getPos4D() const;
 	const Matrix4f getViewMatrix() const override;
-	const Matrix<float, 5> get4DViewMatrix() const;
-	void set4DViewMatrix(const Matrix<float, 5> &matrix);
+	const Matrix<double, 5> get4DViewMatrix() const;
+	void set4DViewMatrix(const Matrix<double, 5> &matrix);
 	void process4DModel(VertexArrayObject & VAO, const Model4D &model) const;
 };
